requirements/requirements.txt
appdirs==1.4.2
Django==1.8.6
django-debug-toolbar==1.6
packaging==16.8
pyparsing==2.1.10
six==1.10.0
sqlparse==0.2.3
